[Fix] betty betty style, and change datatype to size_t

+++ 10-binary_tree_depth.c
@@ -1,14 +1,16 @@
 #include "binary_trees.h"
 /**
- * binary_tree_depth - Function that measures the depth of a node in a binary tree.
+ * binary_tree_depth - Function that measures the
+ * depth of a node in a binary tree.
  *@tree: Is a pointer to the node to measure the depth.
+ * Return: size_t or 0
  */
 size_t binary_tree_depth(const binary_tree_t *tree)
 {
-    int depth = 0;
+	size_t depth = 0;
 
-    if (!tree->parent || !tree)
-	return (0);
-    depth = binary_tree_depth(tree->parent);
-    return (depth + 1);
+	if (!tree->parent || !tree)
+		return (0);
+	depth = binary_tree_depth(tree->parent);
+	return (depth + 1);
 }
